One spec by matcher for property based tests
